// lib/api-zod/src/generated/types/gpsDevelopmentGame.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Belconnen United FC Performance Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { GpsDevelopmentGameMatchType } from './gpsDevelopmentGameMatchType';
import type { GpsDevelopmentGameRoleSource } from './gpsDevelopmentGameRoleSource';
import type { GpsDevelopmentGameValues } from './gpsDevelopmentGameValues';

export interface GpsDevelopmentGame {
  key: string;
  year: string;
  round: string;
  date: string | null;
  opponent: string | null;
  squad: string;
  mins: number | null;
  role: string | null;
  roleSource: GpsDevelopmentGameRoleSource;
  matchType: GpsDevelopmentGameMatchType;
  comparable: boolean;
  qualifies: boolean;
  values: GpsDevelopmentGameValues;
}