// lib/api-zod/src/generated/types/gpsDevelopmentMetricSummary.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Belconnen United FC Performance Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { GpsDevelopmentReference } from './gpsDevelopmentReference';

export interface GpsDevelopmentMetricSummary {
  id: string;
  average: number | null;
  sample: number;
  best: number | null;
  gref: GpsDevelopmentReference;
}